+++ projects/swimlane/ngx-ui/src/lib/components/drawer/drawer.component.spec.ts
@@ -1,5 +1,8 @@
+import { ApplicationRef } from '@angular/core';
+import { BACKSPACE, ESCAPE } from '@angular/cdk/keycodes';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
+import { take } from 'rxjs/operators';
 
 import { DrawerComponent } from './drawer.component';
 import { DrawerDirection } from './drawer-direction.enum';
@@ -54,10 +57,14 @@ describe('DrawerComponent', () => {
   });
 
   describe('onEscapeKey', () => {
-    it('should close', () => {
-      const spy = spyOn(component.close, 'emit');
-      component.onEscapeKey();
-      expect(spy).toHaveBeenCalled();
+    it('should close', done => {
+      // The `close` subject will emit when there is at least 1 observer.
+      const promise = component.close.pipe(take(1)).toPromise();
+      fixture.nativeElement.dispatchEvent(createKeyboardEvent(ESCAPE));
+      promise.then(closed => {
+        expect(closed).toBeTrue();
+        done();
+      });
     });
   });
 
@@ -86,4 +93,33 @@ describe('DrawerComponent', () => {
       expect(component.widthSize).toEqual('100%');
     });
   });
+
+  describe('change detection', () => {
+    it('should not trigger change detection if any key is pressed except `Escape`', () => {
+      const appRef = TestBed.inject(ApplicationRef);
+      const spy = spyOn(appRef, 'tick').and.callThrough();
+      // The `close` subject will emit when there is at least 1 observer.
+      const subscription = component.close.subscribe();
+
+      fixture.nativeElement.dispatchEvent(createKeyboardEvent(BACKSPACE));
+      fixture.nativeElement.dispatchEvent(createKeyboardEvent(BACKSPACE));
+
+      expect(spy).toHaveBeenCalledTimes(0);
+
+      fixture.nativeElement.dispatchEvent(createKeyboardEvent(ESCAPE));
+
+      expect(spy).toHaveBeenCalledTimes(1);
+      subscription.unsubscribe();
+    });
+  });
 });
+
+function createKeyboardEvent(keyCode: number) {
+  const event = new KeyboardEvent('keyup');
+  // There seems to be an issue with TS typings since the type-checker throws:
+  // Object literal may only specify known properties, and 'keyCode' does not exist in type 'KeyboardEventInit'.
+  Object.defineProperty(event, 'keyCode', {
+    get: () => keyCode
+  });
+  return event;
+}